% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.list.R
\name{case.list}
\alias{case.list}
\title{Return a data.frame with all currently available cases}
\usage{
case.list()
}
\description{
\code{case.list} returns a data.frame with all currently available cases
based on the .Rmd files in the /inst directory.
}
