% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r.R
\docType{package}
\name{data.r-package}
\alias{data.r}
\alias{data.r-package}
\title{Data Science with R - Cases}
\description{
This package contains some Data Science cases used within the lectures
of Ronald Hochreiter <ron@hochreiter.net>
}
\details{
For more details visit http://www.data-r.com/
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.data-r.com/}
}

}
