% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.case.R
\name{source.case}
\alias{source.case}
\title{Export the R code of a specific case}
\usage{
source.case(case, output = NA, documentation = 0)
}
\description{
\code{source.case} exports the R code of a given case to output
}
